Refactor and consolidate `pickItem` utilities (#417)
