+++ src/commands/deleteContainerApp/deleteContainerApp.ts
@@ -6,7 +6,7 @@
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import { ContainerAppItem } from "../../tree/ContainerAppItem";
 import { localize } from "../../utils/localize";
-import { pickContainerApp } from "../../utils/pickContainerApp";
+import { pickContainerApp } from "../../utils/pickItem/pickContainerApp";
 
 export async function deleteContainerApp(context: IActionContext, node?: ContainerAppItem): Promise<void> {
     node ??= await pickContainerApp(context, {
