+++ src/commands/editContainerApp.ts
@@ -8,7 +8,7 @@ import type { IActionContext } from "@microsoft/vscode-azext-utils";
 import { ext } from "../extensionVariables";
 import type { ContainerAppItem } from "../tree/ContainerAppItem";
 import { localize } from "../utils/localize";
-import { pickContainerApp } from "../utils/pickContainerApp";
+import { pickContainerApp } from "../utils/pickItem/pickContainerApp";
 
 export async function editContainerApp(context: IActionContext, node?: ContainerAppItem): Promise<void> {
     node ??= await pickContainerApp(context);
