+++ src/commands/ingress/editTargetPort/editTargetPort.ts
@@ -8,7 +8,7 @@ import { IngressEnabledItem } from "../../../tree/configurations/IngressItem";
 import type { ContainerAppItem } from "../../../tree/ContainerAppItem";
 import { createActivityContext } from "../../../utils/activityUtils";
 import { localize } from "../../../utils/localize";
-import { pickContainerApp } from "../../../utils/pickContainerApp";
+import { pickContainerApp } from "../../../utils/pickItem/pickContainerApp";
 import type { IngressContext } from "../IngressContext";
 import { TargetPortInputStep } from "./TargetPortInputStep";
 import { TargetPortUpdateStep } from "./TargetPortUpdateStep";
