+++ src/commands/ingress/toggleIngressVisibility/toggleIngressVisibility.ts
@@ -5,7 +5,7 @@ import type { ContainerAppItem } from "../../../tree/ContainerAppItem";
 import type { IngressEnabledItem } from "../../../tree/configurations/IngressItem";
 import { createActivityContext } from "../../../utils/activityUtils";
 import { localize } from "../../../utils/localize";
-import { pickContainerApp } from "../../../utils/pickContainerApp";
+import { pickContainerApp } from "../../../utils/pickItem/pickContainerApp";
 import type { IngressContext } from "../IngressContext";
 import { ToggleIngressVisibilityStep } from "./ToggleIngressVisibilityStep";
