+++ src/commands/revision/chooseRevisionMode.ts
@@ -10,7 +10,7 @@ import { ext } from "../../extensionVariables";
 import type { ContainerAppItem, ContainerAppModel } from "../../tree/ContainerAppItem";
 import type { RevisionsItem } from "../../tree/revisionManagement/RevisionsItem";
 import { localize } from "../../utils/localize";
-import { pickContainerApp } from "../../utils/pickContainerApp";
+import { pickContainerApp } from "../../utils/pickItem/pickContainerApp";
 import { updateContainerApp } from "../../utils/updateContainerApp";
 
 export async function chooseRevisionMode(context: IActionContext, node?: ContainerAppItem | RevisionsItem): Promise<void> {
