+++ src/commands/secret/addSecret/addSecret.ts
@@ -8,7 +8,7 @@ import { ext } from "../../../extensionVariables";
 import { SecretsItem } from "../../../tree/configurations/secrets/SecretsItem";
 import { createActivityContext } from "../../../utils/activityUtils";
 import { localize } from "../../../utils/localize";
-import { pickContainerApp } from "../../../utils/pickContainerApp";
+import { pickContainerApp } from "../../../utils/pickItem/pickContainerApp";
 import type { ISecretContext } from "../ISecretContext";
 import { SecretCreateStep } from "./SecretCreateStep";
 import { SecretNameStep } from "./SecretNameStep";
